created an dictionaryApi
Created a dictionaryApi const and set up a try and catch object. The endpoint or dictionary api is stored in data with an axios so different browsers can talk to each other and to test it I sent the data to the console. useEffect is used so dictionaryApi method is rendered once because of [] dependency when the react app is rendered. useEffect does not need to be put in the return ().