Add inline style to App.js, Container from @material-ui/core, inline style to the Container, import Container from @material-ui/core
The Container must be placed inside a {} for it to be desentized, or something like that' for it to work. A useState is used with a 'setWord' and a setMeanings, the 'meanings' variable has been sent to console.log to test it. components folder created to house the 'Header.css and the Header.js' files.